+++ ChangeLog.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Show warnings on trait assignment typos like `device.frequency = 5` instead of `device.state.frequency = 5`
-- `state` or `settings` traits can be defined directly in a `labbench.Device` class. They are automatically moved into `state` if they include a getter or setter, or otherwise into `settings`.
+- `state` or `settings` traits can be defined directly in a `labbench.Device` class. Settings are defined as annotations (':') and states are defined with assignment ('=')
 - Add first 40 lines of CommandLineWrapper output to debug logs
 - Removed logger warnings when calls to CommandLineWrapper.kill() do not kill any process
 - Tightened the message about a pending exception in lb.concurrently
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - text file outputs in relational databases are now encoded as utf-8
 - Removed Trait parameters 'default_value', 'read_only', and 'write_only'; added Trait parameters 'default', 'settable', 'gettable', 'allow', have been added
 - Removed Device methods '__get_state__', '__set_state__'; added methods '__get_command__'
+- Support for updating default values of settings in subclasses as annotations
 ### Removed
 -